requirements.txt
python-dotenv==0.15.0
pandas==1.2.4
boto3==1.17.54